fix: id 전략 변경 및 int -> Integer 래퍼로 변경

+++ src/main/java/peermarket/peershop/controller/dto/ItemListDto.java
@@ -17,8 +17,8 @@ public class ItemListDto {
     private Long id;
     private String itemName;
     private String imgUrl;
-    private int price;
-    private int stockQuantity;
+    private Long price;
+    private Integer stockQuantity;
     private LocalDateTime createdAt;
 
     public ItemListDto(Item item) {
